Refactored SDK download logic from git-util into separate package. Added support to remove files from sdk based on globs. wip
