+++ packages/git-util/src/index.ts
@@ -1,97 +1,20 @@
 import oldFs from 'fs';
-import { join, relative } from 'path';
+import { relative } from 'path';
 
-import AdmZip from 'adm-zip';
-import { move } from 'fs-extra';
-import globby from 'globby';
-import { clone, push, add, listFiles, commit } from 'isomorphic-git';
+import { clone, push, add, commit } from 'isomorphic-git';
 
+import {
+  getAllStageableFilepathsInRepo,
+  migrateSdkIntoLocalRepo,
+} from '@openapi-platform/download-util';
+import { deletePaths, makeTempDir } from '@openapi-platform/download-util/lib/file';
 import { GitInfo } from '@openapi-platform/model';
-import { downloadToPath, deletePaths, makeTempDir } from './file/index';
-
-// Note: dir = directory
-
-export async function getAllStageableFilepathsInRepo(repoDir: string) {
-  return await globby([join(repoDir, '**'), join(repoDir, '**', '.*')], {
-    gitignore: true,
-  });
-}
-
-export async function getAllFilepathsInDir(dir: string) {
-  return await globby([join(dir, '**'), join(dir, '**', '.*')]);
-}
-
-export async function moveFilesIntoLocalRepo(repoDir, sdkDir) {
-  // Unfortunately you have to move each file individual (to knowledge)
-  const paths = await getAllFilepathsInDir(sdkDir);
-  for (const path of paths) {
-    const fromPath = path;
-    const relativePath = relative(sdkDir, path);
-    const toPath = join(repoDir, relativePath);
-    await move(fromPath, toPath, { overwrite: true });
-  }
-}
-
-async function deleteAllFilesInLocalRepo(dir) {
-  const filePaths = await listFiles({ fs: oldFs, dir });
-  const fullFilePaths = filePaths.map(path => join(dir, path));
-  await deletePaths(fullFilePaths);
-}
-
-/**
- * SDKs are downloaded as a zip file so they need to be extracted.
- * Use this method to do so.
- * @param archiveFilePath The archive file path
- * @param extractToDir Where you want the archive files to be extracted to
- */
-async function extractSdkArchiveFileToDir(extractToDir: string, archiveFilePath: string) {
-  const zip = new AdmZip(archiveFilePath);
-  await zip.extractAllTo(extractToDir, true);
-}
-
-/**
- * Tries to put the files of a remotely stored sdk ZIP file into a locally stored
- * repository.
- */
-export async function migrateSdkIntoLocalRepo(
-  repoDir: string,
-  remoteSdkUrl: string,
-  options,
-) {
-  const { logger } = options;
-  logger.verbose(`Deleting all files ${repoDir}`);
-  await deleteAllFilesInLocalRepo(repoDir);
-  /*
-   * We make folders for each step of the process,
-   * one for downloading, one for unzipping.
-   */
-  const downloadDir = await makeTempDir('download');
-  try {
-    const sdkArchiveFilePath = join(downloadDir, 'sdk.zip');
-    await downloadToPath(sdkArchiveFilePath, remoteSdkUrl);
-    const sdkDir = await makeTempDir('sdk');
-    try {
-      logger.verbose(`Extracting ${sdkArchiveFilePath} to ${sdkDir}`);
-      await extractSdkArchiveFileToDir(sdkDir, sdkArchiveFilePath);
-
-      logger.verbose(`Moving files from ${sdkDir} to ${repoDir}`);
-      await moveFilesIntoLocalRepo(repoDir, sdkDir);
-    } catch (err) {
-      throw err;
-    } finally {
-      await deletePaths([sdkDir]);
-    }
-  } catch (err) {
-    throw err;
-  } finally {
-    await deletePaths([downloadDir]);
-  }
-}
 
 export async function updateRepoWithNewSdk(
   gitInfo: GitInfo,
   remoteSdkUrl: string,
   options,
+  fileCleaningGlobs = [],
 ) {
   // TODO: Rather than taking a logger, just provide callbacks
   const { logger } = options;
@@ -116,7 +39,7 @@ export async function updateRepoWithNewSdk(
       ...gitInfo.auth,
     });
 
-    await migrateSdkIntoLocalRepo(repoDir, remoteSdkUrl, options);
+    await migrateSdkIntoLocalRepo(repoDir, remoteSdkUrl, fileCleaningGlobs, options);
     const addedPaths = await getAllStageableFilepathsInRepo(repoDir);
     logger.verbose(`Staging ${addedPaths.length} paths`);
     for (const addedPath of addedPaths) {
